Made ability to post and finished the project for now..
Created ability to register and create a new user in the database with avatar username and display name that also has the ability to post tweets with text and optionally with also an image or a gif. To do this the user has to click the register button on the top right, fill out the form and after that a new button will appear for creating posts. With this I end this project for the time being.