Fixed issue 1
There appears to be 2 issues which were not noticed previously, the first one is seen when making a tweet, the tweet is not using the latest registered user but the previous one. The other is that when making multiple posts with images the new image is the one that is only shown, the older ones dissapear.